requirements.txt
discord.py==1.2.5
ipython==7.10.2
